added cnfcertjob and controllerhelper pkgs
