+++ controllers/controllerhelper/definitions.go
@@ -1,0 +1,26 @@
+package controllerhelper
+
+import (
+	"k8s.io/apimachinery/pkg/runtime"
+	"sigs.k8s.io/controller-runtime/pkg/client"
+)
+
+const (
+	CnfCertPodNamePrefix = "cnf-job-run"
+
+	CnfCertSuiteBaseFolder      = "/cnf-certsuite"
+	CnfCnfCertSuiteConfigFolder = CnfCertSuiteBaseFolder + "/config/suite"
+	CnfPreflightConfigFolder    = CnfCertSuiteBaseFolder + "/config/preflight"
+	CnfCertSuiteResultsFolder   = CnfCertSuiteBaseFolder + "/results"
+
+	CnfCertSuiteConfigFilePath    = CnfCnfCertSuiteConfigFolder + "/tnf_config.yaml"
+	PreflightDockerConfigFilePath = CnfPreflightConfigFolder + "/preflight_dockerconfig.json"
+
+	SideCarResultsFolderEnvVar = "TNF_RESULTS_FOLDER"
+)
+
+// CnfCertificationSuiteRunReconciler reconciles a CnfCertificationSuiteRun object
+type CnfCertificationSuiteRunReconciler struct {
+	client.Client
+	Scheme *runtime.Scheme
+}
